Add Authorization.functions to User.functions
